renamed search controller into search worker
