+++ include/bitbishop/interface/search_worker.hpp
